:art: user와 auth를 디렉토리 분리 #68
