+++ src/features/auth/types.d.ts
@@ -1,0 +1,4 @@
+type Account = {
+  id: string;
+  password: string;
+};
