+++ src/models/user/api.ts
@@ -1,0 +1,9 @@
+import { client } from '@/shared/api';
+
+export const getUser = async () => {
+  return client.get<User>(`/user`).then((response) => {
+    const { data } = response;
+    console.debug('user:', data);
+    return data;
+  });
+};
