+++ src/models/user/constants.ts
@@ -1,0 +1,1 @@
+export const queryKey = ['user'];
