+++ src/models/user/index.ts
@@ -1,0 +1,7 @@
+import { queryKey } from './constants';
+
+import { queryClient } from '@/shared/api';
+
+export * from './hooks';
+
+export const removeUserQuery = () => queryClient.removeQueries({ queryKey });
