+++ src/models/user/types.d.ts
@@ -14,8 +14,3 @@ type User = Profile & {
   level: number;
   userCode: string;
 };
-
-type Account = {
-  id: string;
-  password: string;
-};
