+++ src/pages/collection/ui/page.tsx
@@ -10,8 +10,8 @@ import * as styles from './page.css';
 
 import SettingTextButton from '@/components/button/SettingTextButton';
 
-import { useUserInfo } from '@/features/user';
 import { refetchCollection } from '@/models/collection';
+import { useUserInfo } from '@/models/user';
 
 const SelectedSkin = () => {
   const skin = useSelectedSkinUrl();
