+++ src/pages/home/index.tsx
@@ -8,7 +8,7 @@ import ColoredIconButton from '@/components/button/ColoredIconButton';
 import UserDashboard from '@/pages/home/UserDashboard';
 
 import { ROUTE } from '@/constants/routes';
-import { useUserInfo } from '@/features/user';
+import { useUserInfo } from '@/models/user';
 
 const Home = () => {
   const { user } = useUserInfo();
