+++ src/pages/landing/index.tsx
@@ -5,7 +5,7 @@ import * as styles from './index.css';
 import ColoredButton from '@/components/button/ColoredButton/index';
 
 import { ROUTE } from '@/constants/routes';
-import { useGuestLogin } from '@/features/user';
+import { useGuestLogin } from '@/features/auth';
 
 const Landing = () => {
   const navigate = useNavigate();
