+++ src/pages/settings/index.tsx
@@ -4,7 +4,7 @@ import * as styles from './index.css';
 import MemberSetting from './member-setting';
 
 import { ROUTE } from '@/constants/routes';
-import { useUserStatus } from '@/features/user';
+import { useUserStatus } from '@/models/user';
 
 const Settings = () => {
   const { isMember } = useUserStatus();
